Plotting codes and tif figure update.
MATLAB code for plotting updated to remove the *2 factors in the indeces when operating on image matrices. The factors were included due to the previous code running on a laptop for which the screen had 200% zoom, resulting in 2 times the number of actual pixels using getframe compared to the nominal size when setting the figures' positions.